+++ tests/testthat/test_run_model.R
@@ -118,11 +118,7 @@ test_that(
       X2 = s4
     )
     expect_identical(
-      run_model(mod1, mod2, cost = x, effect = y),
-      run_models(mod1, mod2, cost = x, effect = y)
-    )
-    expect_identical(
-      run_models(mod1, mod2, init = 1:0, cost = x, effect = y),
+      run_models(mod1, mod2, init = c(1000L, 0L), cost = x, effect = y),
       run_models(mod1, mod2, cost = x, effect = y)
     )
     expect_identical(
@@ -132,20 +128,20 @@ test_that(
     expect_output(
       str(run_models(mod1, mod2, cost = x, effect = y)),
       '2 obs. of  5 variables:
- $ x           : num  309 934
- $ y           : num  283 891
+ $ x           : num  309300 933900
+ $ y           : num  283300 891300
  $ .model_names: chr  "A" "B"
- $ .cost       : num  309 934
- $ .effect     : num  283 891',
+ $ .cost       : num  309300 933900
+ $ .effect     : num  283300 891300',
       fixed = TRUE
     )
     expect_output(
       str(summary(run_models(mod1, mod2, cost = x, effect = y))),
       "2 obs. of  5 variables:
-  ..$ x      : num [1:2] 309 934
-  ..$ y      : num [1:2] 283 891
-  ..$ .cost  : num [1:2] 0 625
-  ..$ .effect: num [1:2] 0 608
+  ..$ x      : num [1:2] 309300 933900
+  ..$ y      : num [1:2] 283300 891300
+  ..$ .cost  : num [1:2] 0 624600
+  ..$ .effect: num [1:2] 0 608000
   ..$ .icer  : num [1:2] -Inf 1.03",
       fixed = TRUE
     )